+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG7C499D5W472GMTRAXR3N2J"
+        "01KGCTCBDGWMAFCMPVD70120Q5"
     )
     r"""The unique identifier of the data source"""